fix: Change GameCenterKit's isAuthenticated computation
Changed the computation of the `isAuthenticated` property in GameCenterKit to remove the `lazy` attribute and use a standard computed property. The previous `lazy` attribute was causing unexpected behaviour when evaluating the authentication status of the local player. This change ensures that the `isAuthenticated` property accurately reflects the authentication status of the local player.